add implementation of generateProgressDialog to SyncCapableCommCareActivity
